Some changes from CLustered Environment. - Added Testing exec part of install. - fixed some warning errors. - fixed how file is generated in the test as creating file on pfs directly is expensive. -

+++ adapter/test/posix/posix_adapter_rs_test.cpp
@@ -221,8 +221,8 @@ TEST_CASE("BatchedReadStrideDynamicRSRangeSmall",
     int fd = open(info.existing_file.c_str(), O_RDWR);
     REQUIRE(fd != -1);
     for (size_t i = 0; i < info.num_iterations; ++i) {
-      size_t offset = abs(((i * rand_r(&info.offset_seed)) % info.stride_size) %
-                          (info.total_size - info.small_max));
+      size_t offset = abs((int)(((i * rand_r(&info.offset_seed)) % info.stride_size) %
+                          (info.total_size - info.small_max)));
       size_t status = lseek(fd, offset, SEEK_SET);
       REQUIRE(status == offset);
       size_t request_size =
@@ -250,8 +250,8 @@ TEST_CASE("BatchedUpdateStrideDynamicRSRangeSmall",
     int fd = open(info.existing_file.c_str(), O_RDWR);
     REQUIRE(fd != -1);
     for (size_t i = 0; i < info.num_iterations; ++i) {
-      size_t offset = abs(((i * rand_r(&info.offset_seed)) % info.stride_size) %
-                          (info.total_size - info.small_max));
+      size_t offset = abs((int)(((i * rand_r(&info.offset_seed)) % info.stride_size) %
+                          (info.total_size - info.small_max)));
       size_t status = lseek(fd, offset, SEEK_SET);
       REQUIRE(status == offset);
       size_t request_size =
@@ -632,8 +632,8 @@ TEST_CASE("BatchedReadStrideDynamicRSRangeMedium",
     int fd = open(info.existing_file.c_str(), O_RDWR);
     REQUIRE(fd != -1);
     for (size_t i = 0; i < info.num_iterations; ++i) {
-      size_t offset = abs(((i * rand_r(&info.offset_seed)) % info.stride_size) %
-                          (info.total_size - info.medium_max));
+      size_t offset = abs((int)(((i * rand_r(&info.offset_seed)) % info.stride_size) %
+                          (info.total_size - info.medium_max)));
       size_t status = lseek(fd, offset, SEEK_SET);
       REQUIRE(status == offset);
       size_t request_size =
@@ -661,8 +661,8 @@ TEST_CASE("BatchedUpdateStrideDynamicRSRangeMedium",
     int fd = open(info.existing_file.c_str(), O_RDWR);
     REQUIRE(fd != -1);
     for (size_t i = 0; i < info.num_iterations; ++i) {
-      size_t offset = abs(((i * rand_r(&info.offset_seed)) % info.stride_size) %
-                          (info.total_size - info.medium_max));
+      size_t offset = abs((int)(((i * rand_r(&info.offset_seed)) % info.stride_size) %
+                          (info.total_size - info.medium_max)));
       size_t status = lseek(fd, offset, SEEK_SET);
       REQUIRE(status == offset);
       size_t request_size =
@@ -1044,8 +1044,8 @@ TEST_CASE("BatchedReadStrideDynamicRSRangeLarge",
     int fd = open(info.existing_file.c_str(), O_RDWR);
     REQUIRE(fd != -1);
     for (size_t i = 0; i < info.num_iterations; ++i) {
-      size_t offset = abs(((i * rand_r(&info.offset_seed)) % info.stride_size) %
-                          (info.total_size - info.large_max));
+      size_t offset = abs((int)(((i * rand_r(&info.offset_seed)) % info.stride_size) %
+                          (info.total_size - info.large_max)));
       size_t status = lseek(fd, offset, SEEK_SET);
       REQUIRE(status == offset);
       size_t request_size =
@@ -1073,8 +1073,8 @@ TEST_CASE("BatchedUpdateStrideDynamicRSRangeLarge",
     int fd = open(info.existing_file.c_str(), O_RDWR);
     REQUIRE(fd != -1);
     for (size_t i = 0; i < info.num_iterations; ++i) {
-      size_t offset = abs(((i * rand_r(&info.offset_seed)) % info.stride_size) %
-                          (info.total_size - info.large_max));
+      size_t offset = abs((int)(((i * rand_r(&info.offset_seed)) % info.stride_size) %
+                          (info.total_size - info.large_max)));
       size_t status = lseek(fd, offset, SEEK_SET);
       REQUIRE(status == offset);
       size_t request_size =
